Changed imports in tests
