+++ components/org.wso2.identity.webhook.common.event.handler/src/test/resources/testng.xml
@@ -20,7 +20,7 @@
     <test name="login-event-hook-handler-tests" preserve-order="true" parallel="false">
         <classes>
             <class name="org.wso2.identity.webhook.common.event.handler.EventConfigManagerTest"/>
-            <class name="org.wso2.identity.webhook.common.event.handler.util.EventHookHandlerUtilsTest"/>
+            <class name="org.wso2.identity.webhook.common.event.handler.util.EventHookHandlerInternalUtilsTest"/>
             <class name="org.wso2.identity.webhook.common.event.handler.LoginEventHookHandlerTest"/>
             <class name="org.wso2.identity.webhook.common.event.handler.PayloadBuilderFactoryTest"/>
         </classes>
